[diffusion] reuse native FlowUniPC for wan s2v
